ManipulateFiles/requirements.txt
PyPDF2
fpdf
numpy
plotly
Pillow
opencv-python
pytesseract

